feat: implement generate-todo-from-convos script
