+++ ToDo.yaml
@@ -91,7 +91,7 @@ todos:
     status: erledigt
   - id: todo-31
     beschreibung: "generate-todo-from-convos.ts implementieren, um aus conversations.json ToDos abzuleiten"
-    status: offen
+    status: erledigt
   - id: todo-32
     beschreibung: "Kollisionsenergie und Gasbildung in Binärsimulation implementieren"
     status: offen
